ats-backend/resume-parser/requirements.txt
pdf2image
pytesseract
pdfplumber
spacy
nltk
Pillow
